move back rule extraction in app
